add support to delete notes

+++ backend/src/main/java/com/nocountry/courses/service/impl/NoteServiceImpl.java
@@ -80,6 +80,9 @@ public class NoteServiceImpl implements INoteService {
 
     @Override
     public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException(messenger.getMessage(RESOURCE_NOT_FOUND.name(), null, Locale.getDefault()));
+        }
         repository.deleteById(id);
     }
